refactor(output): swap date and name order in market holiday list
The output format for the `make_market_holidays_list` utility has been updated.
The order of the holiday name and date in the output file has been swapped to
present the date first, followed by the holiday name. This change aims to improve
consistency or utility for downstream parsing.

Identified changes:
- Changed the order of arguments to `std::format` from `holiday.first, holiday.second` to `holiday.second, holiday.first`.
- Modified the format string in `std::format` from `"{}\t{:%Y-%m-%d}\n"` to `"{:%Y-%m-%d}\t{}\n"`.
- As a result, the output format has changed from `Holiday Name <TAB> YYYY-MM-DD` to `YYYY-MM-DD <TAB> Holiday Name`.